// packages/neko-types/src/types/asset/entity.d.ts
export type FilePurpose = 'main' | 'thumbnail' | 'preview' | 'texture' | 'reference' | 'source';
/** Media file type */
export type AssetMediaType = 'video' | 'audio' | 'image' | 'sequence' | 'text';